feat: implement MCP tool grouping and search functionality
