+++ backend/docs/swagger.json
@@ -6518,6 +6518,30 @@
                 }
             }
         },
+        "/settings/mcp-policy": {
+            "get": {
+                "security": [
+                    {
+                        "BearerAuth": []
+                    }
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "settings"
+                ],
+                "summary": "查询 MCP 工具运行策略",
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope"
+                        }
+                    }
+                }
+            }
+        },
         "/settings/model-option-policy": {
             "get": {
                 "security": [
@@ -11706,7 +11730,7 @@
                 },
                 "selectedToolIDs": {
                     "type": "array",
-                    "maxItems": 32,
+                    "maxItems": 128,
                     "items": {
                         "type": "integer"
                     }
