+++ backend/internal/application/settings/seed.go
@@ -164,6 +164,7 @@ func defaultSettings() []domainsettings.SystemSetting {
 		{Namespace: "mcp", Key: "mcp_tool_timeout_seconds", Value: "10", ValueType: "int", Description: "MCP Tool Call 超时(秒)"},
 		{Namespace: "mcp", Key: "mcp_tool_retry_count", Value: "0", ValueType: "int", Description: "MCP Tool Call 重试次数"},
 		{Namespace: "mcp", Key: "mcp_max_concurrent_calls", Value: "8", ValueType: "int", Description: "MCP Tool Call 并发上限"},
+		{Namespace: "mcp", Key: "mcp_max_selected_tools_per_message", Value: "32", ValueType: "int", Description: "单次消息最多可选择的 MCP 工具数量"},
 		{Namespace: "mcp", Key: "mcp_max_llm_calls_per_run", Value: "5", ValueType: "int", Description: "单次 MCP 工具运行最大 LLM 请求次数（最小 2，首次请求 + 工具后续请求 + 最终总结）"},
 		{Namespace: "mcp", Key: "mcp_max_tool_calls_per_run", Value: "8", ValueType: "int", Description: "单次 MCP 工具运行最大 MCP Tool Call 次数"},
