+++ backend/internal/application/settings/service_test.go
@@ -135,6 +135,18 @@ func TestValidateModelOptionPolicySettings(t *testing.T) {
 	}
 }
 
+func TestValidateMCPSelectedToolsSetting(t *testing.T) {
+	if err := validatePatchItem(PatchItem{Namespace: "mcp", Key: "mcp_max_selected_tools_per_message", Value: "32"}); err != nil {
+		t.Fatalf("expected selected tool limit to pass, got %v", err)
+	}
+	if err := validatePatchItem(PatchItem{Namespace: "mcp", Key: "mcp_max_selected_tools_per_message", Value: "0"}); err == nil {
+		t.Fatal("expected zero selected tool limit to fail")
+	}
+	if err := validatePatchItem(PatchItem{Namespace: "mcp", Key: "mcp_max_selected_tools_per_message", Value: "129"}); err == nil {
+		t.Fatal("expected selected tool limit above safe maximum to fail")
+	}
+}
+
 func TestValidateFullContextLimitsAllowUnlimitedValues(t *testing.T) {
 	cases := []PatchItem{
 		{Namespace: "file", Key: "full_context_limit_enabled", Value: "true"},
