+++ backend/internal/transport/http/conversation/dto_request.go
@@ -84,7 +84,7 @@ type SendMessageRequest struct {
 	Options               map[string]interface{} `json:"options"`
 	ClientRunID           string                 `json:"clientRunID" binding:"omitempty,max=64"`
 	FileIDs               []string               `json:"fileIDs" binding:"max=20"`
-	SelectedToolIDs       []uint                 `json:"selectedToolIDs" binding:"max=32"`
+	SelectedToolIDs       []uint                 `json:"selectedToolIDs" binding:"max=128"`
 	ParentMessagePublicID string                 `json:"parentMessagePublicID" binding:"omitempty,max=32"`
 	SourceMessagePublicID string                 `json:"sourceMessagePublicID" binding:"omitempty,max=32"`
 	BranchReason          string                 `json:"branchReason" binding:"omitempty,oneof=default retry edit"`
