+++ backend/internal/transport/http/conversation/handler.go
@@ -111,6 +111,9 @@ func mapStreamError(err error) streamError {
 	case errors.Is(err, appconversation.ErrTooManyMessageFiles):
 		status = http.StatusBadRequest
 		message = "too many files in one message"
+	case errors.Is(err, appconversation.ErrTooManySelectedTools):
+		status = http.StatusBadRequest
+		message = "too many selected tools"
 	case errors.Is(err, appconversation.ErrFileProcessingNotReady):
 		status = http.StatusBadRequest
 		message = "file processing not ready"
