+++ backend/internal/transport/http/settings/router.go
@@ -8,6 +8,7 @@ func (m *Module) RegisterPublicRoutes(api *gin.RouterGroup) {
 
 func (m *Module) RegisterRoutes(api *gin.RouterGroup) {
 	api.GET("/settings/model-option-policy", m.Handler.GetModelOptionPolicy)
+	api.GET("/settings/mcp-policy", m.Handler.GetMCPPolicy)
 }
 
 // RegisterAdminRoutes 注册 settings 管理路由（由管理员中间件保护）。
