+++ frontend/features/chat/components/app-chat-area.tsx
@@ -182,6 +182,7 @@ export function AppChatArea() {
     showTokenUsage,
     showBillingCost,
     modelOptionPolicy,
+    mcpMaxSelectedTools,
     selectedPlatformModelName,
     setSelectedPlatformModelName,
   } = useChatModelOptions({
@@ -209,6 +210,15 @@ export function AppChatArea() {
   const [toolsLoading, setToolsLoading] = React.useState(true);
   const [selectedToolIDs, setSelectedToolIDs] = React.useState<number[]>([]);
   const initializedOptionsModelRef = React.useRef("");
+
+  React.useEffect(() => {
+    setSelectedToolIDs((current) => {
+      if (current.length <= mcpMaxSelectedTools) {
+        return current;
+      }
+      return current.slice(0, mcpMaxSelectedTools);
+    });
+  }, [mcpMaxSelectedTools]);
 
   React.useEffect(() => {
     const platformModelName = selectedModel?.platformModelName.trim() || "";
@@ -551,6 +561,7 @@ export function AppChatArea() {
     selectedPlatformModelName,
     availableTools,
     selectedToolIDs,
+    maxSelectedTools: mcpMaxSelectedTools,
     toolsLoading,
     options: effectiveOptions,
     defaultOptions: selectedModelDefaultOptions,
