+++ frontend/i18n/messages/en-US/chat.json
@@ -288,6 +288,14 @@
     "noToolDescription": "No description",
     "viewToolDescription": "View tool description",
     "noMatchingTools": "No matching tools",
+    "mcpUnknownServer": "Unnamed server",
+    "mcpServerToolCount": "{selected} / {total} selected",
+    "mcpSelectServerTools": "Select all",
+    "mcpClearServerTools": "Clear",
+    "mcpToggleServerTools": "Select tools from {server}",
+    "mcpServerLimitHint": "Max {limit}",
+    "mcpToolLimitTitle": "Tool limit reached",
+    "mcpToolLimitDescription": "A message can use up to {limit} MCP tools. Clear some tools before selecting more.",
     "pauseGeneration": "Pause generation",
     "voiceInput": "Voice input",
     "cancelVoiceInput": "Cancel voice input",
