+++ frontend/i18n/messages/zh-CN/chat.json
@@ -288,6 +288,14 @@
     "noToolDescription": "暂无简介",
     "viewToolDescription": "查看工具简介",
     "noMatchingTools": "没有匹配的工具",
+    "mcpUnknownServer": "未命名服务",
+    "mcpServerToolCount": "已选 {selected} / {total}",
+    "mcpSelectServerTools": "全选",
+    "mcpClearServerTools": "取消",
+    "mcpToggleServerTools": "选择 {server} 的工具",
+    "mcpServerLimitHint": "最多 {limit} 个",
+    "mcpToolLimitTitle": "工具数量已达上限",
+    "mcpToolLimitDescription": "单次消息最多选择 {limit} 个 MCP 工具，请减少选择后再继续。",
     "pauseGeneration": "暂停生成",
     "voiceInput": "语音输入",
     "cancelVoiceInput": "取消语音输入",
