+++ frontend/shared/api/settings.ts
@@ -8,6 +8,10 @@ type ModelOptionPolicyResponse = {
   nativeToolAllowedTypesJSON: string;
 };
 
+export type MCPPolicy = {
+  maxSelectedToolsPerMessage: number;
+};
+
 export async function getModelOptionPolicy(accessToken: string): Promise<ModelOptionPolicy> {
   const data = await authedRequest<ModelOptionPolicyResponse>(
     "/api/v1/settings/model-option-policy",
@@ -19,5 +23,16 @@ export async function getModelOptionPolicy(accessToken: string): Promise<ModelOp
     allowedPathsJSON: data.allowedPathsJSON,
     deniedPathsJSON: data.deniedPathsJSON,
     nativeToolAllowedTypesJSON: data.nativeToolAllowedTypesJSON,
+  };
+}
+
+export async function getMCPPolicy(accessToken: string): Promise<MCPPolicy> {
+  const data = await authedRequest<MCPPolicy>(
+    "/api/v1/settings/mcp-policy",
+    { accessToken },
+    true,
+  );
+  return {
+    maxSelectedToolsPerMessage: data.maxSelectedToolsPerMessage,
   };
 }
